--- main.go
@@ -1,13 +1,0 @@
-package main
-
-import (
-	"os"
-
-	"github.com/uitml/frink/cmd"
-)
-
-func main() {
-	if err := cmd.Execute(); err != nil {
-		os.Exit(1)
-	}
-}
